Updated failng execution tests: semanticTests/viaYul.

+++ test/libsolidity/semanticTests/various/crazy_elementary_typenames_on_stack.sol
@@ -1,10 +1,10 @@
 contract C {
     function f() public returns (uint256 r) {
-        uint;
-        uint;
-        uint;
-        uint;
-        int x = -7;
+        uint256;
+        uint256;
+        uint256;
+        uint256;
+        int256 x = -7;
         return uint256(x);
     }
 }
@@ -13,4 +13,4 @@ contract C {
 // compileViaYul: also
 // compileToEwasm: also
 // ----
-// f() -> -7
+// f() -> 0x00fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff9
